# doc/specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Store Daily QA System API",
    "description": "API for store staff daily QA system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/chat/message": {
      "post": {
        "tags": ["chat"],
        "summary": "Send Chat Message",
        "description": "AIチャットメッセージ送信",
        "operationId": "send_chat_message_api_chat_message_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatMessageRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat message sent successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatMessageResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/chat/session/save": {
      "post": {
        "tags": ["chat"],
        "summary": "Save Chat Session",
        "description": "チャットセッション保存",
        "operationId": "save_chat_session_api_chat_session_save_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat session saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionSaveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/chat/session": {
      "get": {
        "tags": ["chat"],
        "summary": "Get Chat Session By Survey",
        "description": "サーベイIDからチャットセッション取得",
        "operationId": "get_chat_session_by_survey_api_chat_session_get",
        "parameters": [
          {
            "name": "survey_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Survey Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat session retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatSessionResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "404": {
            "description": "Chat session not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorResponse" }
        },
        "type": "object",
        "required": ["error"],
        "title": "APIError",
        "description": "API エラーレスポンス"
      },
...
      "AgentType": {
        "type": "string",
        "enum": ["hearing", "coaching", "analysis", "daily_summary"],
        "title": "AgentType",
        "description": "AIエージェントタイプ"
      },
...
      "ChatMessageRequest": {
        "properties": {
          "agent_type": { "$ref": "#/components/schemas/AgentType" },
          "prompt": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Prompt",
            "description": "プロンプト"
          },
          "session_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Session Id",
            "description": "セッションID"
          },
          "survey_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Survey Id",
            "description": "サーベイ回答ID"
          }
        },
        "type": "object",
        "required": ["agent_type", "prompt", "session_id", "survey_id"],
        "title": "ChatMessageRequest",
        "description": "AIチャットメッセージリクエスト"
      },
      "ChatMessageResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response",
            "description": "AI応答"
          },
          "agent_type": { "$ref": "#/components/schemas/AgentType" },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "セッションID"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id",
            "description": "アクターID"
          }
        },
        "type": "object",
        "required": ["response", "agent_type", "session_id", "actor_id"],
        "title": "ChatMessageResponse",
        "description": "AIチャットメッセージレスポンス"
      },
      "ChatSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "セッションID"
          },
          "survey_id": {
            "type": "string",
            "title": "Survey Id",
            "description": "サーベイID"
          },
          "user_cd": {
            "type": "string",
            "title": "User Cd",
            "description": "ユーザーコード"
          },
          "str_cd": {
            "type": "string",
            "title": "Str Cd",
            "description": "店舗コード"
          },
          "messages": {
            "items": { "$ref": "#/components/schemas/MessageHistoryItem" },
            "type": "array",
            "title": "Messages",
            "description": "会話履歴"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "survey_id",
          "user_cd",
          "str_cd",
          "messages",
          "created_at"
        ],
        "title": "ChatSessionResponse",
        "description": "チャットセッション詳細レスポンス"
      },
      "ChatSessionSaveRequest": {
        "properties": {
          "survey_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Survey Id",
            "description": "サーベイID"
          },
          "session_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Session Id",
            "description": "セッションID"
          }
        },
        "type": "object",
        "required": ["survey_id", "session_id"],
        "title": "ChatSessionSaveRequest",
        "description": "チャットセッション保存リクエスト"
      },
      "ChatSessionSaveResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "保存結果メッセージ"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "ChatSessionSaveResponse",
        "description": "チャットセッション保存レスポンス"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "エラーコード"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "エラーメッセージ"
          },
          "details": {
            "anyOf": [{}, { "type": "null" }],
            "title": "Details",
            "description": "エラー詳細"
          }
        },
        "type": "object",
        "required": ["code", "message"],
        "title": "ErrorResponse",
        "description": "エラーレスポンス"
      },
...
      "MessageHistoryItem": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "メッセージの役割 (user/assistant)"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "メッセージ内容"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "メッセージ送信時刻"
          },
          "agent_type": {
            "anyOf": [
              { "$ref": "#/components/schemas/AgentType" },
              { "type": "null" }
            ],
            "description": "AIエージェントタイプ（assistantの場合）"
          }
        },
        "type": "object",
        "required": ["role", "content", "timestamp"],
        "title": "MessageHistoryItem",
        "description": "メッセージ履歴項目"
      },
...
    }
  }
}